Fetch test JSON paths in a more reliable fashion
Several of the unit tests were broken, either by (1) updates to how
cargo works, or (2) the previous patch where I added a top-level
workspace. In any case, `file!` doesn't quite do what we want, because
it only contains a relative path. I've switched to
`std::env::current_dir`, which is used for this purpose by my own
`cli_test_dir` crate and by @burntshushi's popular `xsv` and `ripgrep`
crates, and which seems to be the most popular way to do this.