pact_consumer: Move doctest to tests.rs temporarily
The original version of this test was wrapped in `#[test]`, which meant
that the code was almost completely ignored--it wasn't run, and almost
no compile-time checks were applied. Because of this, the example wasn't
runnable.

So I made the doctest runnable, and fixed all the compilation errors,
and replaced the dummy HTTP client with `reqwest`, a popular lightweight
synchronous wrapper around `hyper` that's great for examples.

Unfortunately, this ran into a possible cargo linking issue, described
here: https://github.com/rust-lang/cargo/issues/4567 So I've temporarily
moved the doctest to tests/tests.rs until upstream has a chance to
figure out what's going on with `cdylib` in doctests.

The next goal will be to simplify the test code here as much as
possible, removing `s!` and other things. But we needed a solid base
with tests to get started.