Replace `s!` with `Into<String>`
The `s!` macro isn't fully idiomatic Rust. A more natural way of
handling these cases is to build APIs that take a parameter satisfying
`Into<String>`, which we can convert into an owned string using
`.into()`.

Note that in some cases, we really want a custom `Into<_>` which allows
JSON, regular expressions, and all the richness that we find in the Ruby
and JavaScript APIs. Soon!

There are two cases we haven't fixed yet, because they take an argument
which contains an _embedded_ `String`: `headers` and `body`.