[BUG] pact_matching: Parse JSON paths with `_`
It appears that most JSON path engines accept identifiers with
underscores. (For an example, see http://jsonpath.com/.) However, query
param matching was failing in `pact_matching` when the query parameters
returned an underscore. This affects many other kinds of matching as
well.

This patch adds support for `_` in JSON paths, and it includes various
test cases I wrote while tracking down the bug.

An alternative to fixing this would be modify `pact_consumer` to use
bracket notation for identifiers with underscores. But it's probably
better to fix the underlying parsing in any case.

Do other Pact implementations handle expressions of the form
`$.query.user_id`? If not, is this by design, or is this a bug?